refactor(news): react-query 중복 개선
